Time: 81 ms (5.03%), Space: 12.7 MB (5.03%) - LeetHub

+++ 174-dungeon-game/174-dungeon-game.cpp
@@ -1,4 +1,4 @@
-typedef tuple<int, int, int> tupl;
+typedef tuple<int, short, short> tupl;
 
 class Solution {
 public:
